A little main menu work and started fixing the collision system
The collision system is still broken, don't be fooled . It's very much still broken.

I removed that placeholder mm.png and actually added a proper main menu back ground. This new main menu also has me adding a couple new functions to the renderer so that I can just make a call to set the zoom to max, or min in this example. I've also updated a couple of the sprites to get ready for the new system where I'll be doing allot of refactoring on the renderer for the caves and such.