Lots of application cleanup and presentation fixes
I did allot of mutation on the renderer to allow it to the a global object rather than creating a new renderer for each scene that uses one. I've also set it up to allow the developer to use the ResourcePacks.
I also added in some dithering to the minimap, which has been a planned feature for some time now. And I think it looks really nice.
I also got the icon working in app by modifying the olcPixelGameEngine.h Platform object.
I've also locked the zoom resolution so the user cannot use the frame decrease to move faster. I may add this feature back in once I fix the movement.