Maybe fixed collision system
I think I've got the collision fixed. It's still a little finicky with the swimming but it's a heck of a lot better than it was before. I've also fixed some things where I had to convert to "screen space" by dividing most things by SPRITE_SCALE. There are still a few remnants of this left behind, but most have been removed.