Added items.h and json.hpp
These two new items are to account for the design philosophy of allowing the user to modify the game. once this is actually implemented the development of features that use these mods will slow down, however adding these new features will be much faster.

The new Item struct will be replacing any definition of an item or even standard object definition. Items will instead be defined as a type. For example an item may be defined as "Food" or as a "Crafting table" or perhaps a "Pickaxe" etc.